requirements.txt
adversarial-robustness-toolbox==1.14.1
asttokens==2.2.1
backcall==0.2.0
bidict==0.22.1
biopython==1.81
biosppy==1.0.0
blinker==1.6.2
certifi==2023.5.7
cffi==1.15.1
charset-normalizer==3.1.0
click==8.1.3
colorama==0.4.6
comm==0.1.3
contourpy==1.0.7
cycler==0.11.0
debugpy==1.6.7
decorator==5.1.1
executing==1.2.0
Flask==2.3.2
fonttools==4.39.4
h5py==3.8.0
idna==3.4
ipykernel==6.23.1
ipython==8.13.2
ipywidgets==8.0.6
itsdangerous==2.1.2
jedi==0.18.2
Jinja2==3.1.2
joblib==1.2.0
jupyter_client==8.2.0
jupyter_core==5.3.0
jupyterlab-widgets==3.0.7
kiwisolver==1.4.4
MarkupSafe==2.1.2
matplotlib==3.7.1
matplotlib-inline==0.1.6
mne==1.4.0
nest-asyncio==1.5.6
nglview==3.0.4
numpy==1.24.3
opencv-python==4.7.0.72
packaging==23.1
pandas==2.0.1
parso==0.8.3
pickleshare==0.7.5
Pillow==9.5.0
platformdirs==3.5.1
pooch==1.7.0
prompt-toolkit==3.0.38
psutil==5.9.5
pure-eval==0.2.2
pycparser==2.21
pyEDFlib==0.1.32
Pygments==2.15.1
pyparsing==3.0.9
python-dateutil==2.8.2
pytz==2023.3
pywin32==306
pywin32-ctypes==0.2.0
pywinpty==2.0.10
pyzmq==25.0.2
requests==2.31.0
scikit-learn==1.1.3
scipy==1.10.1
shortuuid==1.0.11
six==1.16.0
soundfile==0.12.1
stack-data==0.6.2
tensorflow-estimator==2.12.0
tensorflow-intel==2.12.0
tensorflow-io-gcs-filesystem==0.31.0
threadpoolctl==3.1.0
tornado==6.3.2
tqdm==4.65.0
traitlets==5.9.0
tzdata==2023.3
urllib3==2.0.2
wcwidth==0.2.6
Werkzeug==2.3.4
wfdb==4.1.1
widgetsnbextension==4.0.7
WTForms==3.0.1
